Trait tuxedo_parachain_core::ParachainConstraintChecker

pub trait ParachainConstraintChecker: ConstraintChecker {
    // Required method
    fn is_parachain(&self) -> bool;
}

A way for the relay chain validators to determine whether a particular parachain extrinsic is the parachain inherent and whether the parachain inherent data can be extracted from it.

Required Methods

fn is_parachain(&self) -> bool

Object Safety

This trait is not object safe.