Trait tuxedo_parachain_core::SetRelayParentNumberStorage

pub trait SetRelayParentNumberStorage {
    // Required method
    fn set(new_parent_number: u32);
}

An abstraction over setting the ambiently available relay parent block number. This allows it to be mocked during tests and require actual externalities.

Required Methods

fn set(new_parent_number: u32)

Object Safety

This trait is not object safe.